UserCheck server-puolelle, supplier-ryhmän tarkistus, shopid ei kovakoodattu
