+++ imports/plugins/custom/olga/server/methods/supplyContracts.app-test.js
@@ -8,7 +8,7 @@ import { resetDatabase } from 'meteor/xolvio:cleaner';
 import { $ } from 'meteor/jquery';
 import { Reaction } from "/server/api";
 import { Roles } from "meteor/alanning:roles";
-import { UserChecks } from "../../lib/userChecks";
+import { UserChecks } from "../helpers/userChecks";
 import { addProduct } from "/server/imports/fixtures/products";
 import Fixtures from "/server/imports/fixtures";
 import { Products, Orders } from "/lib/collections";
